feat: MainContent 컴포넌트를 서버 컴포넌트로 변경
