+++ src/app/components/Loader/Loader.module.scss
@@ -1,3 +1,5 @@
+@use "/src/styles/variables" as var;
+
 .loader {
   width: 100%;
   height: 100%;
@@ -23,7 +25,7 @@
 .loading:before,
 .loading:after,
 .loading {
-  background: #076fe5;
+  background: var.$primary;
   -webkit-animation: loading-keys-app-loading 0.8s infinite ease-in-out;
   animation: loading-keys-app-loading 0.8s infinite ease-in-out;
   width: 13.6px;
@@ -45,13 +47,13 @@
   80%,
   100% {
     opacity: 0.75;
-    box-shadow: 0 0 #076fe5;
+    box-shadow: 0 0 var.$primary;
     height: 32px;
   }
 
   40% {
     opacity: 1;
-    box-shadow: 0 -8px #076fe5;
+    box-shadow: 0 -8px var.$primary;
     height: 40px;
   }
 }
@@ -61,13 +63,13 @@
   80%,
   100% {
     opacity: 0.75;
-    box-shadow: 0 0 #076fe5;
+    box-shadow: 0 0 var.$primary;
     height: 32px;
   }
 
   40% {
     opacity: 1;
-    box-shadow: 0 -8px #076fe5;
+    box-shadow: 0 -8px var.$primary;
     height: 40px;
   }
 }
