+++ src/app/folder/[[...id]]/OptionList/OptionList.module.scss
@@ -1,0 +1,7 @@
+.optionListContainer {
+  display: flex;
+  align-items: flex-start;
+  list-style: none;
+  padding: 0px;
+  gap: 12px;
+}
