+++ src/app/shared/page.module.scss
@@ -12,6 +12,7 @@
   margin: 2.5rem auto 6.25rem auto;
   display: flex;
   flex-direction: column;
+  padding-bottom: 100px;
 }
 
 .searchBarWrapper {
